tests/wpt/tests/html/interaction/focus/focusgroup/tentative/grid-navigation/does-not-wrap-or-flow.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Simple case that validated we don't wrap or flow when these values aren't provided.</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://open-ui.org/components/scoped-focusgroup.explainer/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/shadow-dom/focus-navigation/resources/focus-utils.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<table focusgroup=grid>
  <tr>
    <td id=r1c1 tabindex=0>r1c1</td>
    <td id=r1c2 tabindex=0>r1c2</td>
    <td id=r1c3 tabindex=0>r1c3</td>
  </tr>
  <tr>
    <td id=r2c1 tabindex=0>r2c1</td>
    <td id=r2c2 tabindex=0>r2c2</td>
    <td id=r2c3 tabindex=0>r2c3</td>
  </tr>
</table>

<script>

  promise_test(async t => {
    var r1c3 = document.getElementById("r1c3");

    await focusAndSendDirectionalInput(r1c3, kRight);
    assert_equals(document.activeElement, r1c3);
  }, "Navigating right should not move the focus when it is at the last column and the focusgroup doesn't wrap or flow.");

  promise_test(async t => {
    var r2c1 = document.getElementById("r2c1");

    await focusAndSendDirectionalInput(r2c1, kDown);
    assert_equals(document.activeElement, r2c1);
  }, "Navigating down should not move the focus when it is at the last row and the focusgroup doesn't wrap or flow.");

  promise_test(async t => {
    var r1c1 = document.getElementById("r1c1");

    await focusAndSendDirectionalInput(r1c1, kLeft);
    assert_equals(document.activeElement, r1c1);
  }, "Navigating left should not move the focus when it is at the first column and the focusgroup doesn't wrap or flow.");

  promise_test(async t => {
    var r1c1 = document.getElementById("r1c1");

    await focusAndSendDirectionalInput(r1c1, kUp);
    assert_equals(document.activeElement, r1c1);
  }, "Navigating up should not move the focus when it is at the first row and the focusgroup doesn't wrap or flow.");

</script>
